Add EBS volume attachment
A volume of type `ebs` can have an `attachment` attribute that
defines the tags of an external EBS volume that will be searched
and if found, attached to the instance.

After attachment, it behaves as the existing EBS volume, except
that the `mount` attribute is now optional, so the volume can be
attached and left alone. Formatting and mounting could then be
done by an init script, for example.

This also brings in a new "uevent listener" whose purpose is to act
as a very simple udev, only for newly attached EBS volumes. It adds
symlinks for the user-defined device names (such as /dev/sda) to
the actual nvme device names.