Configure networking on early boot
This replaces kernel PNP DHCP, which does not always configure the
primary interface to be eth0.

This starts network initialization as the first step of the init
process after creating the tokio runtime. It uses netlink to
discover interfaces and brings up the first by index after
filtering out lo, veth* and other interfaces. It configures the
interface with DHCP, then checks IMDS to verify the correct
primary interface. If it differs, it is reconfigured and then
persisted in a file to be loaded on the next boot.